+++ src/utils/constants.ts
@@ -1,1 +1,7 @@
 export const apiPath = '/api/v1';
+
+export const LOGGER_TYPE = {
+  ERROR: 'error',
+  WARNING: 'warn',
+  INFO: 'info'
+};
