+++ src/utils/serverLogger.ts
@@ -1,3 +1,4 @@
 export const serverLogger = (type, message) => {
+  if (process.env.NODE_ENV === 'test') return;
   console.log(`[${type}] `, message);
 };
